Added generic search field to user document
Beside the current username and real user name, this change introduces a generic field that can be mapped to any LDAP attribute. This new field, named tag, is invisible to the final user but is used in the users search functionality. For instance, suppose the LDAP attribute 'color' is mapped to 'tag'. After that, by typing 'blue' in the search field, one could find all users whose 'color' LDAP attribute is blue.